# Swapi/swagger.json
{
 "swagger":"2.0",
  "info": {
    "title": "StarWars",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "tags": [
    {
      "name": "Peoples"
    },
    {
      "name": "Films"
    },
    {
      "name": "Planets"
    },
    {
      "name": "Species"
    },
    {
      "name": "Starships"
    },
    {
      "name": "Transports"
    },
    {
      "name": "Vehicles"
    }
  ],
  "paths": {
    "/people": {
      "get": {
        "tags": [
          "Peoples"
        ],
        "summary": "Get All",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Peoples"
        ],
        "summary": "Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Antonin Hausseguy",
                  "gender": "female",
                  "skin_color": "white",
                  "hair_color": "green",
                  "height": "130",
                  "eye_color": "red",
                  "mass": "69",
                  "homeworld": 11,
                  "birth_year": "unknown"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/people/{id}": {
      "get": {
        "tags": [
          "Peoples"
        ],
        "summary": "Get One",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Peoples"
        ],
        "summary": "Update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "name": "Antonin Hausseguy II"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "parameters": [
        {
          "schema": {
            "type": "number"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "delete": {
        "summary": "Delete",
        "operationId": "",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Peoples"
        ]
      }
    },
    "/films": {
      "get": {
        "summary": "Get All",
        "operationId": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Films"
        ]
      },
      "post": {
        "tags": [
          "Films"
        ],
        "summary": "Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "title": "Two Girls One Cup",
                  "episode_id": 4,
                  "starships": [
                    1,
                    7,
                    9,
                    15
                  ],
                  "vehicles": [
                    8,
                    6,
                    9,
                    11
                  ],
                  "planets": [
                    1
                  ],
                  "producer": "Antonin, Noe",
                  "director": "Thibault",
                  "release_date": "2201-05-25T00:00:00.000Z",
                  "opening_crawl": "Eating icecream",
                  "characters": [
                    1,
                    2
                  ],
                  "species": [
                    1,
                    2
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/films/{id}": {
      "get": {
        "tags": [
          "Films"
        ],
        "summary": "Get One",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Films"
        ],
        "summary": "Update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "title": "Oops mauvais titre",
                  "starships": [
                    10,
                    20,
                    30,
                    40,
                    50
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "parameters": [
        {
          "schema": {
            "type": "number"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "delete": {
        "summary": "Delete",
        "operationId": "",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Films"
        ]
      }
    },
    "/planets": {
      "get": {
        "tags": [
          "Planets"
        ],
        "summary": "Get All",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Planets"
        ],
        "summary": "Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "climate": "humid",
                  "surface_water": "5",
                  "name": "Pluto",
                  "diameter": "555",
                  "rotation_period": "1555",
                  "terrain": "water",
                  "gravity": "500",
                  "orbital_period": "999",
                  "population": "1"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/planets/{id}": {
      "get": {
        "tags": [
          "Planets"
        ],
        "summary": "Get One",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Planets"
        ],
        "summary": "Update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "population": "200000"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "parameters": [
        {
          "schema": {
            "type": "number"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "delete": {
        "summary": "Delete",
        "operationId": "",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Planets"
        ]
      }
    },
    "/species": {
      "get": {
        "summary": "Get All",
        "operationId": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Species"
        ]
      },
      "post": {
        "tags": [
          "Species"
        ],
        "summary": "Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "classification": "lapin",
                  "name": "Lapin cretin",
                  "designation": "lapin",
                  "eye_colors": "red",
                  "people": [
                    1
                  ],
                  "skin_colors": "white",
                  "language": "laplap",
                  "hair_colors": "bold",
                  "homeworld": 98,
                  "average_lifespan": "1",
                  "average_height": "20"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/species/{id}": {
      "get": {
        "tags": [
          "Species"
        ],
        "summary": "Get One",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Species"
        ],
        "summary": "Update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "eye_colors": "diamond",
                  "people": [
                    1,
                    4,
                    5,
                    6,
                    12,
                    14,
                    81,
                    82
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "parameters": [
        {
          "schema": {
            "type": "number"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "delete": {
        "summary": "Delete",
        "operationId": "",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Species"
        ]
      }
    },
    "/starships": {
      "get": {
        "tags": [
          "Starships"
        ],
        "summary": "Get All",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Starships"
        ],
        "summary": "Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "pilots": [
                    2
                  ],
                  "MGLT": "89",
                  "starship_class": "ScoobyDoo Car",
                  "hyperdrive_rating": "1.0"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/starships/{id}": {
      "get": {
        "tags": [
          "Starships"
        ],
        "summary": "Get One",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Starships"
        ],
        "summary": "Update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "hyperdrive_rating": "99.0",
                  "pilots": [
                    1
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "parameters": [
        {
          "schema": {
            "type": "number"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "delete": {
        "summary": "Delete",
        "operationId": "",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Starships"
        ]
      }
    },
    "/vehicles": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Get All",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "vehicle_class": "trotinnete",
                  "pilots": [
                    1
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/vehicles/{id}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Get One",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "pilots": [
                    1,
                    4
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "operationId": "patch-vehicles-id"
      },
      "parameters": [
        {
          "schema": {
            "type": "number"
          },
          "name": "id",
          "in": "path",
          "required": true
        }
      ],
      "delete": {
        "summary": "Delete",
        "operationId": "",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Vehicles"
        ]
      }
    }
  },
  "components": {
    "responses": {}
  }
}